Set Up of all the Main Menus and In-games Menus (except from the Inventory)
Changed the texts sizes, set up buttons events and organized everything on the Menus.
It was not set up only the Scene transitions, like on the Play button to call the Game / "Level 1" scene, or the Restart Level or the Main Menu button to go back to the Main Menu from the Game scene.